+++ R/fillHoles.r
@@ -10,6 +10,8 @@
 #'
 #' @example man/examples/ex_GVector.r
 #'
+#' @seealso [terra::fillHoles()], **GRASS** manual page for module `v.fill.holes` (see `grassHelp("v.fill.holes")`)
+#'
 #' @aliases fillHoles
 #' @rdname fillHoles
 #' @exportMethod fillHoles