+++ R/longlat.r
@@ -1,8 +1,10 @@
 #' Create longitude/latitude rasters
 #' 
-#' @description `longlat()` creates two rasters, one with cell values equal to the longitude of the cell centers, and one with cell values equal to the latitude of the cell centers. Values will be in decimal degrees, regardless of the projection of the raster. If you want projected coordinates, use [init()].
+#' @description `longlat()` creates two rasters, one with cell values equal to the longitude of the cell centers, and one with cell values equal to the latitude of the cell centers.
 #' 
 #' @param x A `GRaster`.
+#'
+#' @param degrees Logical: If `TRUE` (default), coordinate values of cells will be in degrees. If `FALSE`, and `x` is in a projected coordinate reference system, values will represent coordinates in map units (usually meters). Values will always be in degrees when the coordinate reference system is unprojected (e.g., WGS84, NAD83, etc.).
 #' 
 #' @returns A `GRaster` stack.
 #'
@@ -16,18 +18,18 @@
 methods::setMethod(
     f = "longlat",
     signature(x = "GRaster"),
-    function(x) {
+    function(x, degrees = TRUE) {
 
     .locationRestore(x)
     .region(x)
-    srcs <- .longlat(x)
+    srcs <- .longlat(x, degrees = degrees)
    .makeGRaster(srcs, c("longitude", "latitude"))
 
     } # EOF
 )
 
 #' @noRd
-.longlat <- function(x) {
+.longlat <- function(x, degrees) {
 
     if (inherits(x, "GRaster")) {
         src <- sources(x)
@@ -36,19 +38,33 @@ methods::setMethod(
     }
 
     srcs <- .makeSourceName(c("longlat_long", "longlat_lat"), "raster")
-    rgrass::execGRASS(
-        cmd = "r.latlong",
-        input = src,
-        output = srcs[1L],
-        flags = c("l", .quiet(), "overwrite")
-    )
+    if (degrees) {
 
-    rgrass::execGRASS(
-        cmd = "r.latlong",
-        input = src,
-        output = srcs[2L],
-        flags = c(.quiet(), "overwrite")
-    )
+        rgrass::execGRASS(
+            cmd = "r.latlong",
+            input = src,
+            output = srcs[1L],
+            flags = c("l", .quiet(), "overwrite")
+        )
+
+        rgrass::execGRASS(
+            cmd = "r.latlong",
+            input = src,
+            output = srcs[2L],
+            flags = c(.quiet(), "overwrite")
+        )
+
+    } else {
+    
+        # longitude
+        ex <- paste0(srcs[1L], " = col(", src, ")")
+        rgrass::execGRASS("r.mapcalc", expression = ex, flags = c(.quiet(), "overwrite"))
+    
+        # latitude
+        ex <- paste0(srcs[2L], " = row(", src, ")")
+        rgrass::execGRASS("r.mapcalc", expression = ex, flags = c(.quiet(), "overwrite"))
+    
+    }
     srcs
 
 }