+++ man/examples/ex_fillNAs.r
@@ -10,11 +10,11 @@ madElev <- fastData("madElev")
 elev <- fast(madElev)
 
 ### Fill NAs:
-biline <- interp <- fillNAs(elev)
-bicube <- interp <- fillNAs(elev, method = "bicubic")
-rst <- interp <- fillNAs(elev, method = "rst")
+bilinear <- fillNAs(elev)
+bicubic <- fillNAs(elev, method = "bicubic")
+rst <- fillNAs(elev, method = "rst")
 
-maps <- c(elev, biline, bicube, rst)
+maps <- c(elev, bilinear, bicubic, rst)
 names(maps) <- c("original", "bilinear", "bicubic", "RST")
 plot(maps)
 
@@ -22,7 +22,7 @@ plot(maps)
 constrained <- fillNAs(elev, min = 0)
 
 # Compare unconstrained and constrained:
-minmax(biline)
+minmax(bilinear)
 minmax(constrained)
 
 ### Interpolate to only first 10 cells away from non-NA cells: